requirements.txt
streamlit>=1.28.0
openai>=1.3.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
icalendar>=5.0.11
streamlit-mermaid>=0.1.0